Import commit; HW from d84d8cf5
This commit is the import of the OS from https://github.com/Digilent/Petalinux-Cora-Z7-07S for the Vivado 2020.1 updated version. The previous repository will be kept for older versions, but this branch will be used for new updates moving forward.
Changes from 2017.4:
Changed the U-Boot and Kernel back to Xilinx's
U-Boot's CONFIG_*(not CONFIG_SYS_*) are now handled by petalinux-config